+++ webtarot.nix
@@ -10,17 +10,23 @@ in
 
 rustPlatform.buildRustPackage rec {
   pname = "webtarot";
-  version = "0.1.0";
+  version = "0.2.6";
 
   src = pkgs.fetchFromGitHub {
     owner = "mmai";
     repo = pname;
-    rev = version;
-    sha256 = "117angn2l8w39wzj07xz435mi2xmdh8xywhjm55mavjnj0yc70p0";
+    rev = "v${version}";
+    sha256 = "15fjj7wbvb9i04dclvhm9ynhlnw91wk4yln0ccra5bnpw0jamacr";
   };
   # src = ./.;
 
-  cargoSha256 = "17kmp4p0aqndvp4k4gzrnk7dq7i2iaz3mkpinki83bxac2y6micj";
+  postInstall = ''
+    mkdir -p $out
+    cp -R ./webtarot_client/static $out/front
+    cp ./webtarot_client/dist/*.{css,js,wasm} $out/front
+    '';
+
+  cargoSha256 = "0x6x0hzgn7x2agw3ah3363pklhmiyacjaxk6fqs9dcan5zz7cs0f";
 
   meta = with pkgs.stdenv.lib; {
     description = "A online game of french tarot";