+++ webtarot_protocol/src/game.rs
@@ -105,15 +105,18 @@ impl GameState {
         }
         players.sort_by(|a, b| a.pos.to_n().cmp(&b.pos.to_n()));
         let pos = self.players[&player_id].pos;
+        let mut scores = [0.0; 5];
+        let mut dog = cards::Hand::new();
+        let mut taker_diff = 0.0;
         let deal = match self.deal.deal_state() {
             Some(state) => { // In Playing phase
-                let  scores =  match state.get_deal_result() {
-                    deal::DealResult::Nothing => [0.0; 5],
-                    deal::DealResult::GameOver {points: _, taker_won: _, scores } => scores
+                if let deal::DealResult::GameOver {points: _, taker_diff: diff, scores: lscores } = state.get_deal_result() {
+                     scores = lscores;
+                     taker_diff = diff;
+                     dog = state.dog();
                 };
-                let last_trick = if self.turn == Turn::Intertrick {
+                let last_trick = if self.turn == Turn::Intertrick && !self.was_last_trick() {
                     // intertrick : there is at least a trick done
-                    // (current_trick() returns the new empty one)
                     state.last_trick().unwrap().clone()
                 } else {
                     state.current_trick().clone()
@@ -126,19 +129,25 @@ impl GameState {
                     hand: state.hands()[pos as usize],
                     current: state.next_player(),
                     contract,
+                    king: state.king(),
                     scores,
                     // last_trick: state.tricks.last().unwrap_or(trick::Trick::default()),
                     last_trick,
                     initial_dog,
+                    dog,
+                    taker_diff,
                 }
             },
             None => DealSnapshot { // In bidding phase
                 hand: self.deal.hands()[pos as usize],
                 current: self.deal.next_player(),
                 contract,
+                king: None,
                 scores: [0.0;NB_PLAYERS],
                 last_trick: trick::Trick::default(),
                 initial_dog: cards::Hand::new(),
+                dog,
+                taker_diff,
             }
         };
         GameStateSnapshot {
@@ -193,7 +202,6 @@ impl GameState {
                 }
                 if count == NB_PLAYERS {
                     if self.turn == Turn::Interdeal { // ongoing game
-                        self.next_deal();
                         self.update_turn();
                     } else { // new game
                         self.turn = Turn::Bidding((bid::AuctionState::Bidding, pos::PlayerPos::P0));
@@ -222,9 +230,11 @@ impl GameState {
             ProtocolError::new(ProtocolErrorKind::InternalError, "unknown auction")
         )?;
         let pass_result = auction.pass(pos);
-        if Ok(bid::AuctionState::Over) == pass_result  {
-            self.complete_auction()?;
-        }
+        match pass_result {
+            Ok(bid::AuctionState::Over) => self.complete_auction()?,
+            Ok(bid::AuctionState::Cancelled) => self.next_deal(),
+            _ => ()
+        };
         self.update_turn();
         Ok(())
     }
@@ -254,11 +264,13 @@ impl GameState {
 
     pub fn set_play(&mut self, pid: Uuid, card: cards::Card) -> Result<(), ProtocolError> {
         let pos = self.players.get(&pid).map(|p| p.pos).unwrap();
-        let state = self.deal.deal_state_mut().unwrap();
+        let state = self.deal.deal_state_mut().ok_or(
+            ProtocolError::new(ProtocolErrorKind::InternalError, "Unknown deal state")
+        )?;
         match state.play_card(pos, card)? {
             deal::TrickResult::Nothing => (),
             deal::TrickResult::TrickOver(_winner, deal::DealResult::Nothing) => self.end_trick(),
-            deal::TrickResult::TrickOver(_winner, deal::DealResult::GameOver{points: _, taker_won: _, scores}) => {
+            deal::TrickResult::TrickOver(_winner, deal::DealResult::GameOver{points: _, taker_diff: _, scores}) => {
                 self.scores.push(scores);
                 self.end_last_trick();
             }
@@ -315,11 +327,16 @@ impl GameState {
     }
 
     fn next_deal(&mut self) {
-        let auction = bid::Auction::new(self.first);
         self.first = self.first.next();
+        let auction = bid::Auction::new(self.first);
         self.deal = Deal::Bidding(auction);
     }
 
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub enum PlayEvent {
+    Play( Uuid, cards::Card)
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
@@ -347,6 +364,35 @@ impl GameStateSnapshot {
             _ => None
         }
     }
+
+    pub fn pos_player_name(&self, pos: pos::PlayerPos) -> String {
+        self.players.iter()
+            .find(|p| p.pos == pos)
+            .map(|found| &found.player.nickname)
+            .unwrap() // panic on invalid pos
+            .into()
+    }
+
+    pub fn current_player_name(&self) -> String {
+        let found_name = self.get_playing_pos().map(|pos| {
+            self.pos_player_name(pos)
+        });
+
+        if let Some(name) = found_name {
+            format!("{}", name)
+        } else {
+            "".into()
+        }
+    }
+
+    // pub fn get_playing_player(&self) -> Option<&GamePlayerState> {
+    //     self.get_playing_player().map(|pos| {
+    //         self.players.iter()
+    //             .find(|p| p.pos == pos)
+    //             .map(|found| &found)
+    //     }).flatten()
+    // }
+
 }
 
 impl Default for GameStateSnapshot {
@@ -360,9 +406,12 @@ impl Default for GameStateSnapshot {
                 hand: cards::Hand::new(),
                 current: pos,
                 contract: None,
+                king: None,
                 scores: [0.0;NB_PLAYERS],
                 last_trick: trick::Trick::new(pos),
                 initial_dog: cards::Hand::new(),
+                dog: cards::Hand::new(),
+                taker_diff: 0.0,
             }
         }
     }
@@ -374,11 +423,55 @@ pub struct GameInfo {
     pub join_code: String,
 }
 
+//Used for server diagnostics
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct GameExtendedInfo {
+    pub game: GameInfo,
+    pub players: Vec<Uuid>
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use tarotgame::{deal_seeded_hands, bid, cards};
+    use tarotgame::{deal_hands, deal_seeded_hands, bid, cards};
     // use tarotgame::{bid, cards, pos, deal, trick};
+
+    #[test]
+    fn test_all_pass() {
+        let mut game = GameState::default();
+        let pos0 = game.add_player(PlayerInfo {id: Uuid::new_v4(), nickname: String::from("player0")});
+        let pos1 = game.add_player(PlayerInfo {id: Uuid::new_v4(), nickname: String::from("player1")});
+        let pos2 = game.add_player(PlayerInfo {id: Uuid::new_v4(), nickname: String::from("player2")});
+        let pos3 = game.add_player(PlayerInfo {id: Uuid::new_v4(), nickname: String::from("player3")});
+        let pos4 = game.add_player(PlayerInfo {id: Uuid::new_v4(), nickname: String::from("player4")});
+
+        assert_eq!(true, game.is_joinable());
+
+        let id0 = game.player_by_pos(pos0).unwrap().player.id;
+        let id1 = game.player_by_pos(pos1).unwrap().player.id;
+        let id2 = game.player_by_pos(pos2).unwrap().player.id;
+        let id3 = game.player_by_pos(pos3).unwrap().player.id;
+        let id4 = game.player_by_pos(pos4).unwrap().player.id;
+        game.set_player_ready(id0);
+        game.set_player_ready(id1);
+        game.set_player_ready(id2);
+        game.set_player_ready(id3);
+        game.set_player_ready(id4);
+        assert_eq!(false, game.is_joinable());
+        assert_eq!(game.get_turn(), Turn::Bidding((bid::AuctionState::Bidding, pos0)));
+
+        let hands_deal1 = game.deal.hands();
+
+        game.set_pass(id0).unwrap();
+        game.set_pass(id1).unwrap();
+        game.set_pass(id2).unwrap();
+        game.set_pass(id3).unwrap();
+        game.set_pass(id4).unwrap();
+
+        //5 passes : should start a new deal
+        assert_eq!(game.get_turn(), Turn::Bidding((bid::AuctionState::Bidding, pos1)));
+        assert_ne!(game.deal.hands(), hands_deal1);
+    }
 
     #[test]
     fn test_game() {