+++ webtarot_server/src/universe.rs
@@ -7,6 +7,7 @@ use warp::ws;
 
 use crate::game::Game;
 use crate::protocol::{Message, PlayerInfo, ProtocolError, ProtocolErrorKind};
+use crate::protocol::{GameInfo, GameExtendedInfo};
 use crate::utils::generate_join_code;
 
 pub struct UniversePlayerState {
@@ -35,6 +36,26 @@ impl Universe {
                 joinable_games: HashMap::new(),
             })),
         }
+    }
+
+    /// show all the active games
+    pub async fn show_games(self: &Arc<Self>) -> Vec<GameExtendedInfo> {
+        let state = self.state.read().await;
+        let fgames = state.games.iter()
+            .map(|(uuid, g)| {
+                g.game_extended_info()
+            } );
+        futures::future::join_all(fgames).await
+    }
+
+    /// for debug purposes: show all the players connected to the server, except player_id
+    pub async fn show_players(self: &Arc<Self>, player_id: Uuid) -> Vec<Uuid> {
+        let state = self.state.read().await;
+        let uuids:Vec<Uuid> = state.players.keys()
+            .filter(|k| *k != &player_id)
+            .map(|k| *k )
+            .collect();
+        uuids
     }
 
     /// Starts a new game.