+++ Changelog.md
@@ -1,5 +1,42 @@
 # Changelog
 
+## 0.2.6
+
+- keep websocket alive : server sends ping every 55s (default nginx timeout = 60s)
+- show dog at deal end
+- show points margin to contract at deal end
+- extended server status
+- message 'your turn to play!'
+- bidding stops after everybody talked
+- show players names instead of position
+
+## 0.2.5
+
+- keep alive websocket
+
+## 0.2.3
+
+- maj static files
+
+## 0.2.2
+
+- new deal when all players pass
+- scores styling
+
+## 0.2.1
+
+- nix front packaging
+
+## 0.2.0
+
+- sounds
+- error messages
+- fix server crashing when playing at bidding phase
+
+## 0.1.1
+
+- colors
+
 ## 0.1.0
 
 - initial release : basic game is playable